ai_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.13
langgraph==0.0.30
langchain-openai==0.1.1
pydantic==2.6.3
python-dotenv==1.0.1
